# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: my_customTools API
  description: |
    媒体转码工具集后端 API 文档。

    包含视频转码、图片处理、动图转换、文档转换等功能。

    ## 认证方式
    使用 JWT Bearer Token 认证，在请求头中添加：
    ```
    Authorization: Bearer <token>
    ```

    ## 响应格式
    所有接口统一返回格式：
    ```json
    {
      "code": 0,
      "msg": "ok",
      "data": {}
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: 开发环境
  - url: http://production-server/api
    description: 生产环境
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token 认证
  schemas:
...
    ApiResponse:
      type: object
      required:
        - code
        - msg
        - data
      properties:
        code:
          type: integer
          description: 业务状态码，0 表示成功
          example: 0
        msg:
          type: string
          description: 响应消息
          example: "ok"
        data:
          type: object
          description: 响应数据
    PaginatedData:
      type: object
      properties:
        list:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            pageSize:
              type: integer
            totalPages:
              type: integer

    # ========== 用户相关 ==========
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
        refreshToken:
          type: string
    RegisterRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          format: password
          minLength: 6
    ChangePasswordRequest:
      type: object
      required:
        - oldPassword
        - newPassword
      properties:
        oldPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password
          minLength: 6
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string

    # ========== 任务相关 ==========
    TaskStatus:
      type: string
      enum: [waiting, uploading, processing, completed, failed, cancelled]
    TaskType:
      type: string
      enum: [video, img, anim, document]
    BaseTask:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/TaskType'
        userId:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        inputPath:
          type: string
        outputPath:
          type: string
        inputFormat:
          type: string
        outputFormat:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        config:
          type: object
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        errorMsg:
          type: string
    CreateTaskRequest:
      type: object
      required:
        - type
        - fileName
        - inputPath
        - outputFormat
      properties:
        type:
          $ref: '#/components/schemas/TaskType'
        fileName:
          type: string
        fileSize:
          type: integer
        inputPath:
          type: string
        inputFormat:
          type: string
        outputFormat:
          type: string
        config:
          oneOf:
            - $ref: '#/components/schemas/VideoTranscodeConfig'
            - $ref: '#/components/schemas/ImgTranscodeConfig'
            - $ref: '#/components/schemas/AnimTranscodeConfig'

    # ========== 转码配置 ==========
    VideoTranscodeConfig:
      type: object
      properties:
        videoCodec:
          type: string
          enum: [h264, h265, vp9, av1, copy]
        resolution:
          type: string
          example: "1920x1080"
        bitrate:
          type: string
          example: "4M"
        crf:
          type: integer
          minimum: 0
          maximum: 51
        fps:
          type: integer
        audioCodec:
          type: string
          enum: [aac, mp3, opus, copy]
        audioBitrate:
          type: string
          example: "128k"
        startTime:
          type: number
        endTime:
          type: number
        rotate:
          type: integer
          enum: [90, -90, 180]
        hwAccel:
          type: string
          enum: [nvidia, vaapi, videotoolbox]
    ImgTranscodeConfig:
      type: object
      properties:
        outputFormat:
          type: string
          enum: [jpg, jpeg, png, webp, avif, bmp, tiff, ico, heic]
        quality:
          type: integer
          minimum: 1
          maximum: 100
        lossless:
          type: boolean
        resizeMode:
          type: string
          enum: [none, width, height, scale, fixed, crop]
        width:
          type: integer
        height:
          type: integer
        keepAspect:
          type: boolean
        colorSpace:
          type: string
          enum: [original, srgb, rgb, cmyk, grayscale]
        stripMeta:
          type: boolean
        stripICC:
          type: boolean
        rotate:
          type: integer
          enum: [0, 90, 180, 270]
        flipH:
          type: boolean
        flipV:
          type: boolean
        background:
          type: string
        compression:
          $ref: '#/components/schemas/ImageCompressionConfig'
    AnimTranscodeConfig:
      type: object
      properties:
        outputFormat:
          type: string
          enum: [gif, webp, apng]
        fps:
          type: integer
        width:
          type: integer
        height:
          type: integer
        colors:
          type: integer
        quality:
          type: integer
        loop:
          type: integer
        dither:
          type: string
          enum: [bayer, floyd, none]
        optimize:
          type: boolean
        delay:
          type: integer
        startTime:
          type: number
        duration:
          type: number
        imagePaths:
          type: array
          items:
            type: string
    ImageCompressionConfig:
      type: object
      properties:
        mode:
          type: string
          enum: [quality, size, percent]
        targetSize:
          type: integer
          description: 目标大小（字节）
        targetPercent:
          type: integer
          minimum: 1
          maximum: 100
        minQuality:
          type: integer
          minimum: 1
          maximum: 100

    # ========== 文档转换 ==========
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: 用户注册
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '400':
          description: 参数错误或用户已存在
        '403':
          description: 系统已关闭注册功能
  /auth/login:
    post:
      tags: [Auth]
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 用户名或密码错误
  /auth/refresh:
    post:
      tags: [Auth]
      summary: 刷新访问令牌
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          accessToken:
                            type: string
                          refreshToken:
                            type: string
        '401':
          description: 刷新令牌无效
  /auth/me:
    get:
      tags: [Auth]
      summary: 获取当前用户信息
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '404':
          description: 用户不存在
  /auth/password:
    put:
      tags: [Auth]
      summary: 修改密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 密码修改成功
        '400':
          description: 旧密码错误或参数无效
...
  /tasks:
    post:
      tags: [Tasks]
      summary: 创建任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '200':
          description: 创建成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BaseTask'
        '400':
          description: 参数错误

    get:
      tags: [Tasks]
      summary: 查询任务列表
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/TaskType'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedData'
...
  /clean/run:
    post:
      tags: [Clean]
      summary: 执行手动清理
      description: 需要管理员权限
      responses:
        '200':
          description: 清理完成
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        description: 清理统计
        '403':
          description: 权限不足
...
  /clean/tasks:
    post:
      tags: [Clean]
      summary: 清理旧任务
      description: 需要管理员权限
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                maxAgeDays:
                  type: integer
                  default: 7
                  description: 清理多少天前的任务
      responses:
        '200':
          description: 清理成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          tasksRemoved:
                            type: integer
        '403':
          description: 权限不足
  /clean/temp:
    post:
      tags: [Clean]
      summary: 清理临时文件
      description: 需要管理员权限
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                maxAgeHours:
                  type: integer
                  default: 24
                  description: 清理多少小时前的临时文件
      responses:
        '200':
          description: 清理成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          filesRemoved:
                            type: integer
        '403':
          description: 权限不足

  # ==================== Export ====================
...